refactor(logic): single-source economy, setup, turn events, diplomacy, order suggestion
- data: add purchaseLandCost() in riches_prices (15× base; validation + application)
- logic: add build_cost.dart (canAffordBuild, applyBuildCostDeduction) for validators + orders_application
- logic: add orders_application_helpers (isMineralEligibleTile)
- logic: add turn_resolution_events (research/diplomacy/capture/victory event emission)
- logic: add initial_visibility.dart; game_setup delegates visibility there
- logic: add diplomacy_relation_lookup (pairKey, getRelation, upsertRelation, getOverture, getPlayer, costs)
- logic: add order_suggestion_helpers (getProvinceOwnerMap, filterMoveOrdersByDiplomacy)
- BuildOrderValidator and WorkOrderValidator use shared build_cost / purchaseLandCost
- Turn resolver phase switch delegates event emission to turn_resolution_events
- order_suggestion and diplomacy_resolver re-export helpers for backward compatibility
- Tests pass (logic, data, ai).

Made-with: Cursor
